CC-1909: Refactor, consolidate validation of server-specific options in ValidateServerOptions()
